Remove Browser project
