+++ .github/workflows/build-dotnet.yml
@@ -34,7 +34,7 @@ jobs:
       - run: |
           dotnet workload install ios
           dotnet restore
-          dotnet build -c Release /p:IsBrowser=true ./Trinsic
+          dotnet build -c Release -f net6.0 -p:IsBrowser=true ./Trinsic
           dotnet build -c Release
           dotnet test -c Release -v n
         working-directory: ./dotnet
