--- dotnet/Trinsic.Browser/Trinsic.Browser.csproj
@@ -1,20 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk.Razor">
-
-    <PropertyGroup>
-        <TargetFramework>net6.0</TargetFramework>
-        <ImplicitUsings>enable</ImplicitUsings>
-        <Nullable>enable</Nullable>
-        <!-- disable packing until fully tested -->
-        <IsPackable>false</IsPackable>
-        <DefineConstants>$(DefineConstants);__BROWSER__</DefineConstants>
-    </PropertyGroup>
-
-    <ItemGroup>
-      <ProjectReference Include="..\Trinsic\Trinsic.csproj" />
-    </ItemGroup>
-
-    <ItemGroup>
-      <PackageReference Include="Microsoft.AspNetCore.Components.WebAssembly" Version="6.0.3" />
-    </ItemGroup>
-
-</Project>
