+++ dotnet/TrinsicSdk.sln
@@ -7,8 +7,6 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Trinsic", "Trinsic\Trinsic.
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Tests", "Tests\Tests.csproj", "{8132C26B-F80A-49EC-933E-DAA7C77235F0}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Trinsic.Browser", "Trinsic.Browser\Trinsic.Browser.csproj", "{85F0B4B0-17B7-4F9A-BA22-C588EAD1CBE2}"
-EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -46,17 +44,5 @@ Global
 		{8132C26B-F80A-49EC-933E-DAA7C77235F0}.Release|x64.Build.0 = Release|Any CPU
 		{8132C26B-F80A-49EC-933E-DAA7C77235F0}.Release|x86.ActiveCfg = Release|Any CPU
 		{8132C26B-F80A-49EC-933E-DAA7C77235F0}.Release|x86.Build.0 = Release|Any CPU
-		{85F0B4B0-17B7-4F9A-BA22-C588EAD1CBE2}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{85F0B4B0-17B7-4F9A-BA22-C588EAD1CBE2}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{85F0B4B0-17B7-4F9A-BA22-C588EAD1CBE2}.Debug|x64.ActiveCfg = Debug|Any CPU
-		{85F0B4B0-17B7-4F9A-BA22-C588EAD1CBE2}.Debug|x64.Build.0 = Debug|Any CPU
-		{85F0B4B0-17B7-4F9A-BA22-C588EAD1CBE2}.Debug|x86.ActiveCfg = Debug|Any CPU
-		{85F0B4B0-17B7-4F9A-BA22-C588EAD1CBE2}.Debug|x86.Build.0 = Debug|Any CPU
-		{85F0B4B0-17B7-4F9A-BA22-C588EAD1CBE2}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{85F0B4B0-17B7-4F9A-BA22-C588EAD1CBE2}.Release|Any CPU.Build.0 = Release|Any CPU
-		{85F0B4B0-17B7-4F9A-BA22-C588EAD1CBE2}.Release|x64.ActiveCfg = Release|Any CPU
-		{85F0B4B0-17B7-4F9A-BA22-C588EAD1CBE2}.Release|x64.Build.0 = Release|Any CPU
-		{85F0B4B0-17B7-4F9A-BA22-C588EAD1CBE2}.Release|x86.ActiveCfg = Release|Any CPU
-		{85F0B4B0-17B7-4F9A-BA22-C588EAD1CBE2}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 EndGlobal
